backend/pyproject.toml
[project]
name = "dietition"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
  { name = "i-kondou", email = "u573077g@ecs.osaka-u.ac.jp" }
]
requires-python = ">=3.12"

dependencies = [
  "jinja2>=3.1.6",
  "ruff>=0.11.11",
  "python-dotenv>=0.9.0",
  "fastapi>=0.115.12",
  "uvicorn>=0.34.3",
  "python-multipart>=0.0.20",
  "firebase-admin>=6.9.0",
  "google-generativeai>=0.8.5",
  "langchain-google-genai>=2.0.10",
  "pillow>=11.2.1",
  "google-cloud-firestore>=2.21.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
